feat: style Member Progress, Member's tasks & Footer
