--- src/components/Footer/Footer.module.scss
@@ -1,7 +1,0 @@
-@import '../Styleguide/Stileguide.scss';
-
-.navbar {
-  grid-area: f;
-  background-color: $secondaryColor;
-  font-family: $font;
-}
