+++ src/components/common/TableHeader/TableHeader.jsx
@@ -1,7 +1,8 @@
 import React from 'react';
+
 const TableHeader = ({ titleArr }) => {
-  return titleArr.map((key, index) => {
-    return <th key={index}>{key.toUpperCase()}</th>;
+  return titleArr.map((el, index) => {
+    return <th key={index}>{el}</th>;
   });
 };
